refactor(Formula, InputView): 입력 검증 기능 Formula로 이동 및 입력 값 변환 기능 구현
